Try/catch around user defined event handlers
